[ALPS05568686] ccci: porting mt6877 change
porting mt6877 change for S

MTK-Commit-Id: ee2fe422fccf230c072e702726eefdf9b4d717f8

CR-Id: ALPS05568686
Feature: Modem Interface Driver

+++ drivers/misc/mediatek/eccci/hif/ccci_hif_internal.h
@@ -242,7 +242,7 @@ struct dvfs_ref {
 	int c1_freq; /* Cluster 1 */
 	int c2_freq; /* Cluster 2 */
 	int c3_freq; /* Cluster 3 */
-	u8 dram_lvl;
+	int dram_lvl;
 	u8 irq_affinity;
 	u8 task_affinity;
 	u8 rps;
